Rename file di component
