+++ src/modules/timesheet/components/entry/editor/forms/EntryDetailFormView.tsx
@@ -1,6 +1,6 @@
 import { FormMode } from '@generic/types';
 import { Card, CardContent, CardHeader } from '@material-ui/core';
-import { EntryDetailFormProps } from '@timesheet/components/entry/editor/forms/EntryDetailForm';
+import { EntryDetailFormProps } from '@timesheet/components/entry/editor/forms/TimesheetEntryDetailForm';
 import * as React from 'react';
 import { FormattedMessage } from 'react-intl';
 import { Field } from 'redux-form';
