+++ src/modules/timesheet/components/timesheetRouter.tsx
@@ -1,29 +1,29 @@
-import { EntryDetail } from '@timesheet/components/entry/detail/EntryDetail';
-import { EntryList } from '@timesheet/components/entry/list/EntryList';
+import { TimesheetEntryDetail } from '@timesheet/components/entry/detail/TimesheetEntryDetail';
+import { TimesheetEntryList } from '@timesheet/components/entry/list/TimesheetEntryList';
 import * as React from 'react';
 import { Route, RouteComponentProps, Switch } from 'react-router';
-import { ApprovalDetail } from './approval/detail/ApprovalDetail';
-import { ApprovalList } from './approval/list/ApprovalList';
-import EntryEditor from './entry/editor/EntryEditor';
+import { TimesheetApprovalDetail } from './approval/detail/TimesheetApprovalDetail';
+import { TimesheetApprovalList } from './approval/list/TimesheetApprovalList';
+import TimesheetEntryEditor from './entry/editor/TimesheetEntryEditor';
 
 type AllProps 
   = RouteComponentProps;
 
 const listComponent = () => (
-  <EntryList orderBy="uid" direction="descending"/>
+  <TimesheetEntryList orderBy="uid" direction="descending"/>
 );
 const detailComponent = () => (
-  <EntryDetail/>
+  <TimesheetEntryDetail/>
 );
 
 const editorComponent = () => (
-  <EntryEditor/>
+  <TimesheetEntryEditor/>
 );
 const approvalListComponent = () => (
-  <ApprovalList orderBy="uid" direction="descending"/>
+  <TimesheetApprovalList orderBy="uid" direction="descending"/>
 );
 const approvalDetailComponent = () => (
-  <ApprovalDetail/>
+  <TimesheetApprovalDetail/>
 );
 
 export const timesheetRouter: React.SFC<AllProps> = props => (
